requirements.txt
openai
slack_sdk
slack_bolt
python-dotenv
apscheduler
aiohttp 
pytz